Correct KSkillData wire types to native short/UCHAR and keep verified common serializers

+++ KncWX2Server.CSharp14/Protocol/CommonSimpleData.cs
@@ -13,14 +13,8 @@ public sealed class KBuyGPItemInfo
     public bool Serialize(NativePrimitiveSerializer serializer) =>
         new NativeUserClassSerializer(serializer).Put(this, static (ser, value) =>
         {
-            ser.Put(value.ItemId);
-            ser.Put(value.Price);
-            ser.Put(value.PvpPoint);
-            ser.Put(value.PeriodType);
-            ser.Put(value.Quantity);
-            ser.Put(value.Endurance);
-            ser.Put(value.Period);
-            return true;
+            ser.Put(value.ItemId); ser.Put(value.Price); ser.Put(value.PvpPoint); ser.Put(value.PeriodType);
+            ser.Put(value.Quantity); ser.Put(value.Endurance); ser.Put(value.Period); return true;
         });
 
     public static bool TryDeserialize(NativePrimitiveSerializer serializer, out KBuyGPItemInfo value)
@@ -30,16 +24,9 @@ public sealed class KBuyGPItemInfo
         {
             if (!ser.TryGet(out int itemId) || !ser.TryGet(out int price) || !ser.TryGet(out int pvpPoint) ||
                 !ser.TryGet(out sbyte periodType) || !ser.TryGet(out int quantity) ||
-                !ser.TryGet(out short endurance) || !ser.TryGet(out short period))
-                return (false, existing);
-
-            existing.ItemId = itemId;
-            existing.Price = price;
-            existing.PvpPoint = pvpPoint;
-            existing.PeriodType = periodType;
-            existing.Quantity = quantity;
-            existing.Endurance = endurance;
-            existing.Period = period;
+                !ser.TryGet(out short endurance) || !ser.TryGet(out short period)) return (false, existing);
+            existing.ItemId = itemId; existing.Price = price; existing.PvpPoint = pvpPoint; existing.PeriodType = periodType;
+            existing.Quantity = quantity; existing.Endurance = endurance; existing.Period = period;
             return (true, existing);
         });
     }
@@ -56,12 +43,7 @@ public sealed class KStat
     public bool Serialize(NativePrimitiveSerializer serializer) =>
         new NativeUserClassSerializer(serializer).Put(this, static (ser, value) =>
         {
-            ser.Put(value.BaseHp);
-            ser.Put(value.AtkPhysic);
-            ser.Put(value.AtkMagic);
-            ser.Put(value.DefPhysic);
-            ser.Put(value.DefMagic);
-            return true;
+            ser.Put(value.BaseHp); ser.Put(value.AtkPhysic); ser.Put(value.AtkMagic); ser.Put(value.DefPhysic); ser.Put(value.DefMagic); return true;
         });
 
     public static bool TryDeserialize(NativePrimitiveSerializer serializer, out KStat value)
@@ -70,13 +52,8 @@ public sealed class KStat
         return new NativeUserClassSerializer(serializer).TryGet(out value, static (ser, existing) =>
         {
             if (!ser.TryGet(out int hp) || !ser.TryGet(out int atkPhysic) || !ser.TryGet(out int atkMagic) ||
-                !ser.TryGet(out int defPhysic) || !ser.TryGet(out int defMagic))
-                return (false, existing);
-            existing.BaseHp = hp;
-            existing.AtkPhysic = atkPhysic;
-            existing.AtkMagic = atkMagic;
-            existing.DefPhysic = defPhysic;
-            existing.DefMagic = defMagic;
+                !ser.TryGet(out int defPhysic) || !ser.TryGet(out int defMagic)) return (false, existing);
+            existing.BaseHp = hp; existing.AtkPhysic = atkPhysic; existing.AtkMagic = atkMagic; existing.DefPhysic = defPhysic; existing.DefMagic = defMagic;
             return (true, existing);
         });
     }
@@ -84,15 +61,13 @@ public sealed class KStat
 
 public sealed class KSkillData
 {
-    public int SkillId { get; set; }
-    public sbyte SkillLevel { get; set; }
+    public short SkillId { get; set; }
+    public byte SkillLevel { get; set; }
 
     public bool Serialize(NativePrimitiveSerializer serializer) =>
         new NativeUserClassSerializer(serializer).Put(this, static (ser, value) =>
         {
-            ser.Put(value.SkillId);
-            ser.Put(value.SkillLevel);
-            return true;
+            ser.Put(value.SkillId); ser.Put(value.SkillLevel); return true;
         });
 
     public static bool TryDeserialize(NativePrimitiveSerializer serializer, out KSkillData value)
@@ -100,10 +75,8 @@ public sealed class KSkillData
         value = new();
         return new NativeUserClassSerializer(serializer).TryGet(out value, static (ser, existing) =>
         {
-            if (!ser.TryGet(out int skillId) || !ser.TryGet(out sbyte skillLevel))
-                return (false, existing);
-            existing.SkillId = skillId;
-            existing.SkillLevel = skillLevel;
+            if (!ser.TryGet(out short skillId) || !ser.TryGet(out byte skillLevel)) return (false, existing);
+            existing.SkillId = skillId; existing.SkillLevel = skillLevel;
             return (true, existing);
         });
     }
